refactor code into peopleService
